# Optimize Class#new