This is the result of modifying `Class#new` with one of my experiments from #1045 and rendering an intentionally large VDOM in a Clearwater app went from 303ms down to 267ms in Chrome, a time savings of 11.8% — roughly what I saw in those experiments.

After the VM invokes the JIT/optimizer (about 2-3 renders in), rendering without this patch converged on 123ms. With this patch, those same renders hit 101ms. This is 17.9% better, so subsequent renders benefit even more.