@meh 2015-08-08
I don't think this is the right way, I'm pretty sure this optimization can be done directly in the compiler.

In short, not emitting the argument for the splat in the function body.