@meh 2015-08-08
That is if I'm getting the optimization preventing issue correctly. Can you confirm please?