@elia 2015-08-08
Yes, if we manage to fix this at compiler level all splat calls will benefit.

Here's where the `slice.call` happens:
https://github.com/opal/opal/blob/master/lib/opal/nodes/def.rb#L129-L134

And here's the var list (params):
https://github.com/opal/opal/blob/master/lib/opal/nodes/def.rb#L94