@jgaskins 2015-08-08
I wasn't sure if it would actually benefit all calls because I only ran a few benchmarks in 2 browsers, so I figured I'd at least add a PR for code I already wrote that I know does help. :-)