@meh 2015-08-08
I don't think it has any practical difference, it's still just a method call.