@elia 2015-08-08
http://jsperf.com/opal-reassign-splat-args

I tried on the browsers I have available