@meh 2015-08-08
Looks good on here, I'll patch the compiler if nobody wants to take it.