@jgaskins 2015-08-08
@elia Nice, I added runs for Safari 9 and Chrome Canary. It's even more drastic than I thought on some browsers.