@elia 2015-08-08
@meh it's all yours 😁
@jgaskins cool, I tried WebKit nightly but I think it's being reported as safari 8