@vais 2015-08-09
@meh :+1: There is another splat issue that needs to be fixed in the compiler: https://github.com/opal/opal/issues/934 Basically, the code generated by the compiler should call `to_a` on the splat argument. Maybe while you are in there you can fix this one too :sweat_smile: