@meh 2015-08-09
@jgaskins can you try the benchmarks again to make sure I didn't fuck the implementation up please?