@meh 2015-08-09
Cool, thanks for spending time profiling, we really need more of it :panda_face: