@meh 2015-08-09
@jgaskins also if you can test with the latest commit as well, I had to change splat **calls** to make them actually compliant with Ruby, I'm pretty sure the helper will be optimized out, but you never know.