@jgaskins 2015-08-10
@meh Looks good:

```
Opal Default, no args, capture block: 27636548

Optimized JS, no args, capture block: 27814075

Opal Default, two args, no block: 28775224

Optimized JS, two args, no block: 29000361

Opal Default, two args, capture block: 26465777

Optimized JS, two args, capture block: 25358184
```

I ran it a few times. The `Opal Default` version was always a very slight bit lower for the first two pairs, which is probably due to `Opal.splat`, but the difference iss probably less than a GC pause would make.

This is really good work. :+1: